requirements.txt
asgiref==3.8.1
asn1crypto==1.5.1
blinker==1.9.0
cached-property==2.0.1
cffi==1.17.1
click==8.1.8
coincurve==20.0.0
colorama==0.4.6
cryptography==44.0.1
cytoolz==1.0.1
distlib==0.3.8
eth-hash==0.7.1
eth-typing==5.2.0
eth-utils==5.2.0
filelock==3.15.4
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Mail==0.10.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
noise==1.2.2
numpy==2.2.3
phe==1.5.0
pillow==11.1.0
platformdirs==4.2.2
py-ecc==7.0.1
pyarrow==19.0.1
pycparser==2.22
pycryptodome==3.21.0
pylance==0.23.0
pynoise==2.3.0
pynput==1.7.7
pyopencl==2025.1
pyotp==2.9.0
python-gnupg==0.5.4
pytools==2025.1.1
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.5.1
toolz==1.0.0
typing_extensions==4.12.2
tzdata==2024.1
virtualenv==20.26.3
Werkzeug==3.1.3
